refactor: simplified main.dart, moved App logic to app.dart and desktopMain to desktop_main.dart
